Correct resizeGrid function and finish CSS styling.
Simplify resizeGrid function by deleting the 2nd if .. else statement
and leaving only the first one (testing the RegEx), as with the current
way the grid is created, the original version can be recreated without
recurring to createGrid function, adn thus it was redundant.

Added a way to clear the text input camp after user clicks the "create"
button within the resizeGrid function.